requirements.txt
# AgentIndex Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Redis (task queue)
redis==5.2.0
rq==1.16.2

# HTTP / Crawling
httpx==0.28.0
aiohttp==3.11.0

# GitHub API
pygithub==2.5.0

# Ollama (local LLM)
ollama==0.4.0

# Data processing
pydantic==2.10.0

# Scheduling
apscheduler==3.10.4

# Logging
structlog==24.4.0

# Environment
python-dotenv==1.0.1
